Move conditions evaluate
