+++ internal/pkg/service/buffer/condition/evaluate_test.go
@@ -1,4 +1,4 @@
-package model_test
+package condition
 
 import (
 	"testing"
@@ -11,7 +11,7 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/store/model"
 )
 
-func TestImportConditions_Evaluate_Defaults(t *testing.T) {
+func TestEvaluate_DefaultImportConditions(t *testing.T) {
 	t.Parallel()
 
 	now, _ := time.Parse(time.RFC3339, "2010-01-01T01:01:01Z")
@@ -19,10 +19,10 @@ func TestImportConditions_Evaluate_Defaults(t *testing.T) {
 	before20Min := now.Add(-20 * time.Minute)
 
 	// Defaults
-	ic := model.DefaultImportConditions()
+	cn := model.DefaultImportConditions()
 
 	// Defaults not met
-	res, desc := ic.Evaluate(now, before01Min, statistics.Value{
+	res, desc := evaluate(cn, now, before01Min, statistics.Value{
 		RecordsCount: 50,
 		RecordsSize:  1 * datasize.KB,
 	})
@@ -30,7 +30,7 @@ func TestImportConditions_Evaluate_Defaults(t *testing.T) {
 	assert.Equal(t, "no condition met", desc)
 
 	// Default count met
-	res, desc = ic.Evaluate(now, before01Min, statistics.Value{
+	res, desc = evaluate(cn, now, before01Min, statistics.Value{
 		RecordsCount: 20000,
 		RecordsSize:  1 * datasize.MB,
 	})
@@ -38,7 +38,7 @@ func TestImportConditions_Evaluate_Defaults(t *testing.T) {
 	assert.Equal(t, "count threshold met, received: 20000 rows, threshold: 10000 rows", desc)
 
 	// Default size met
-	res, desc = ic.Evaluate(now, before01Min, statistics.Value{
+	res, desc = evaluate(cn, now, before01Min, statistics.Value{
 		RecordsCount: 100,
 		RecordsSize:  10 * datasize.MB,
 	})
@@ -46,7 +46,7 @@ func TestImportConditions_Evaluate_Defaults(t *testing.T) {
 	assert.Equal(t, "size threshold met, received: 10.0 MB, threshold: 5.0 MB", desc)
 
 	// Default time met
-	res, desc = ic.Evaluate(now, before20Min, statistics.Value{
+	res, desc = evaluate(cn, now, before20Min, statistics.Value{
 		RecordsCount: 100,
 		RecordsSize:  1 * datasize.KB,
 	})
@@ -54,7 +54,7 @@ func TestImportConditions_Evaluate_Defaults(t *testing.T) {
 	assert.Equal(t, "time threshold met, opened at: 2010-01-01T00:41:01.000Z, passed: 20m0s threshold: 5m0s", desc)
 }
 
-func TestImportConditions_Evaluate_Custom(t *testing.T) {
+func TestEvaluate(t *testing.T) {
 	t.Parallel()
 
 	now, _ := time.Parse(time.RFC3339, "2010-01-01T01:01:01Z")
@@ -62,13 +62,13 @@ func TestImportConditions_Evaluate_Custom(t *testing.T) {
 	before20Min := now.Add(-20 * time.Minute)
 
 	// Defaults
-	ic := model.Conditions{
+	cn := Conditions{
 		Count: 100,
 		Size:  5 * datasize.MB,
 		Time:  10 * time.Minute,
 	}
 	// Not met
-	res, desc := ic.Evaluate(now, before01Min, statistics.Value{
+	res, desc := evaluate(cn, now, before01Min, statistics.Value{
 		RecordsCount: 50,
 		RecordsSize:  1 * datasize.MB,
 	})
@@ -76,7 +76,7 @@ func TestImportConditions_Evaluate_Custom(t *testing.T) {
 	assert.Equal(t, "no condition met", desc)
 
 	// Count met
-	res, desc = ic.Evaluate(now, before01Min, statistics.Value{
+	res, desc = evaluate(cn, now, before01Min, statistics.Value{
 		RecordsCount: 200,
 		RecordsSize:  1 * datasize.MB,
 	})
@@ -84,7 +84,7 @@ func TestImportConditions_Evaluate_Custom(t *testing.T) {
 	assert.Equal(t, "count threshold met, received: 200 rows, threshold: 100 rows", desc)
 
 	// Size met
-	res, desc = ic.Evaluate(now, before01Min, statistics.Value{
+	res, desc = evaluate(cn, now, before01Min, statistics.Value{
 		RecordsCount: 50,
 		RecordsSize:  10 * datasize.MB,
 	})
@@ -92,7 +92,7 @@ func TestImportConditions_Evaluate_Custom(t *testing.T) {
 	assert.Equal(t, "size threshold met, received: 10.0 MB, threshold: 5.0 MB", desc)
 
 	// Time met
-	res, desc = ic.Evaluate(now, before20Min, statistics.Value{
+	res, desc = evaluate(cn, now, before20Min, statistics.Value{
 		RecordsCount: 50,
 		RecordsSize:  1 * datasize.MB,
 	})
